feat: Update select menu defaultValue when moving books
We use onChange to trigger events, and thus we need the
select menu to display the proper default value in order
to register changes. If the default value is always the
top, it would be impossible to move books into the top
shelf. Note, the BooksAPI does not return the updated book,
and so the shelf on the book object is updated manually.